Restaurants ViewModel recyclerview for restaurants and a lot of stuff

+++ app/src/main/java/com/example/restaurants/models/Restaurant.kt
@@ -4,4 +4,4 @@ data class Restaurant (val id: String,
                        val name: String,
                        val location: Location,
                        val url: String,
-                       val overallRating: Int = 0)
+                       val overallRating: Float = 0.00f)
